Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
psycopg2-binary
pymongo
python-dotenv
web3
twilio
scikit-learn
pandas
numpy
xgboost
joblib
tensorflow
torch
torchvision
torchaudio
torch-geometric